server/requirements.txt
eventlet==0.27.0
Flask-RESTful==0.3.8
Flask==1.1.2
Flask-Sockets==0.2.1
google-cloud-datastore==1.15.0
gunicorn==20.0.4
requests==2.24.0
